reconfiguring OpenEXP to use the MEAN stack
