+++ pages/index.html
@@ -1,8 +1,12 @@
 <!doctype html>
-<html>
+<html ng-app="myApp">
   <head>
     <title>OpenEXP</title>
-    <script type="text/javascript" src="js/jquery.min.js" onload="window.$ = window.jQuery = module.exports;"></script>
+
+    <!-- Javascript -->
+    <!--<script type="text/javascript" src="js/jquery.min.js" onload="window.$ = window.jQuery = module.exports;"></script>-->
+
+    <script type="text/javascript" src="js/jquery.min.js"></script>
     <script src="js/jQueryRotate.js"></script>
     <script src="js/bootstrap.min.js"></script>
     <script src="externals/jspsych/jspsych.js"></script>
@@ -10,17 +14,21 @@
     <script src="externals/jspsych/plugins/jspsych-single-stim.js"></script>
     <script src="http://d3js.org/d3.v3.min.js" charset="utf-8"></script>
     <script src="https://cdn.socket.io/socket.io-1.3.7.js"></script>
+    <script src="https://code.angularjs.org/1.5.0-beta.2/angular.js"></script>
+    <script src="https://code.angularjs.org/1.5.0-beta.2/angular-route.js"></script>
+    <script src="app.js"></script>
+
     <!-- CSS -->
     <link href="css/bootstrap.css" rel="stylesheet">
     <link href="externals/jspsych/css/jspsych.css" rel="stylesheet" type="text/css"></link>
     <link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css" rel="stylesheet" >
-
     <link href="css/custom.css" rel="stylesheet">
-
   </head>
+
   <body>
     <nav class="navbar navbar-inverse">
     <div class="container-fluid">
+
       <!-- Brand and toggle get grouped for better mobile display -->
       <div class="navbar-header">
         <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
@@ -37,13 +45,13 @@
       <!-- Collect the nav links, forms, and other content for toggling -->
       <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
         <ul class="nav navbar-nav">
-          <li class="active"><a href="index.html">OpenEXP<span class="sr-only">(current)</span></a></li>
-          <li><a href="#">EEG</a></li>
-          <li><a href="sandbox.html">Sandbox</a></li>
+          <li class="active"><a href="#/main.html">OpenEXP<span class="sr-only">(current)</span></a></li>
+          <li><a href="#/EEG">EEG</a></li>
+          <li><a href="#/sandbox">Sandbox</a></li>
           <li class="dropdown">
             <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Experiments <span class="caret"></span></a>
             <ul class="dropdown-menu">
-              <li><a href="reactionTimeTask.html">Reaction Time Task</a></li>
+              <li><a href="#/reactionTimeTask">Reaction Time Task</a></li>
               <li><a href="#">Memory Study - Pictures</a></li>
               <li><a href="#">Language Study - Words</a></li>
               <li role="separator" class="divider"></li>
@@ -69,12 +77,17 @@
       </div><!-- /.navbar-collapse -->
     </div><!-- /.container-fluid -->
   </nav>
+
   <div class="row spacer" style="margin-top:100px"></div>
+  <div ng-controller="mainController">
+  <div ng-view></div> <!--this is where we will inject the code-->
+  </div>
+
   <center>
-  <div class="hidden">
-    <div id="fire">
+  <div>
+
       <img id="spinit" src="img/headerlogofront.png" width="300"/>
-    </div>
+
   </div>
   <div class="row spacer" style="margin-top:100px"></div>
   <h3 class="">OpenEXP is an open-source browser-based interface for running experiments and collecting EEG data with OpenBCI.</h4>
@@ -82,12 +95,15 @@
   <!--<button class="btn btn-lg btn-default">Learn More</button>-->
   </center>
   </body>
-
   <script>
+
+/*
+  // fade in the logo
   $(document).ready(function () {
     $('div.hidden').fadeIn(1000).removeClass('hidden');
   });
 
+  // spin the OpenBCI logo!
   var value = 0
   $("#spinit").rotate({
     bind:
@@ -98,6 +114,9 @@
       }
     }
   });
+
+  */
   </script>
 
+</body>
 </html>
