+++ pages/main.html
@@ -1,0 +1,31 @@
+<center>
+<div class="hidden">
+  <div id="fire">
+    <img id="spinit" src="img/headerlogofront.png" width="300"/>
+  </div>
+</div>
+<div class="row spacer" style="margin-top:100px"></div>
+<h3 class="">OpenEXP is an open-source browser-based interface for running experiments and collecting EEG data with OpenBCI.</h4>
+<div class="row spacer" style="margin-top:100px"></div>
+<!--<button class="btn btn-lg btn-default">Learn More</button>-->
+</center>
+</body>
+<script>
+
+// fade in the logo
+$(document).ready(function () {
+  $('div.hidden').fadeIn(1000).removeClass('hidden');
+});
+
+// spin the OpenBCI logo!
+var value = 0
+$("#spinit").rotate({
+  bind:
+  {
+    click: function(){
+      value +=90;
+      $(this).rotate({ animateTo:value})
+    }
+  }
+});
+</script>
