+++ reactionTimeTask.html
@@ -65,7 +65,7 @@
     </div><!-- /.container-fluid -->
   </nav>
   <div class="row spacer" style="margin-top:100px"></div>
-  <div id="jspsych-target"></div>
+  <center><div id="jspsych-target"></div></center>
   </body>
   <script>
